[fix] 07/05 빌드 오류 수정
