+++ app/recruit/[recruitId]/components/ApplyRequirementSection.tsx
@@ -1,4 +1,3 @@
-import clsx from "clsx";
 import RecruitmentDataItem from "./RecruitmentDataItem";
 
 interface ApplyRequirementSectionProps {
