+++ app/recruit/[recruitId]/components/WorkingConditionSection.tsx
@@ -1,4 +1,3 @@
-import clsx from "clsx";
 import RecruitmentDataItem from "./RecruitmentDataItem";
 
 interface WorkingConditionSectionProps {
